+++ src/Traits/HasRoles.php
@@ -219,20 +219,6 @@ trait HasRoles
     }
 
     /**
-     * @deprecated deprecated since version 1.0.1, use hasPermissionTo instead
-     *
-     * Determine if the user may perform the given permission.
-     *
-     * @param Permission $permission
-     *
-     * @return bool
-     */
-    public function hasPermission($permission)
-    {
-        return $this->hasPermissionTo($permission);
-    }
-
-    /**
      * Determine if the user has, via roles, the given permission.
      *
      * @param Permission $permission