+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `laravel-permission` will be documented in this file
 ## 2.0.0
 
 - cache expiration is now configurable and set to 24h by default
+- removed deprecated `hasPermission` method
 
 ## 1.12.0